+++ client/CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 - Changed advanced search date input to accept full dates ([#39](https://github.com/project-lux/lux-frontend/issues/39)).
+- Changed facet dates to accept full dates ([#39](https://github.com/project-lux/lux-frontend/issues/39)).
 
 ### Removed